requirements.txt
# Core
streamlit==1.45.0
python-dotenv==1.1.0

# LLM & Agents
openai==1.78.1
langchain==0.3.25
langchain-openai==0.3.18
langchain-community==0.3.24
langgraph==0.4.1

# RAG & Vector Store
chromadb==1.0.8
sentence-transformers==4.1.0

# OCR & Image Processing
easyocr==1.7.2
Pillow==11.2.1

# Audio Processing
# Using OpenAI's Whisper API instead of local whisper

# Database
sqlalchemy==2.0.41

# Utilities
pydantic==2.11.4
numpy==1.26.4